Tmp11

+++ cmftbl-smt-solver/src/main/kotlin/de/moritzhank/cmftbl/smt/solver/translation/formula/generation/GenerateVarIntroNodes.kt
@@ -12,7 +12,7 @@ internal fun generateVarIntroNodes(
   usedUnboundVars: Set<CCB<*>>,
   evalCtx: EvaluationContext,
   evalTickIndex: Int,
-  evalInterval: Pair<Int, Int>?,
+  evalInterval: Pair<Double, Double>?,
   evalTickPrecondition: EvaluationTickPrecondition?
 ): Pair<List<VarIntroNode>, EvaluationContext> {
   val varIntroNodes = mutableListOf<VarIntroNode>()
